+++ pages/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from 'next'
 
 import Banner from '../components/Banner'
 import Footer from '../components/Footer'
-import ProductList from '../components/ProductList'
+import ProductList from './products/ProductList'
 import styles from '../styles/Home.module.css'
 
 const Home: NextPage = () => {
@@ -12,8 +12,6 @@ const Home: NextPage = () => {
       <Banner/>
       <ProductList/>
       <Footer/>
-      
-      
 </div>
 </div>
   )
